# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VLTONE Platform API
  version: 1.0.0
  description: Source-of-truth contract shared by the public site, admin site, and desktop application.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/bug-reports:
    post:
      operationId: createBugReport
      security: [{ webCookie: [], csrf: [] }]
      requestBody:
        required: true
        content: { multipart/form-data: { schema: { $ref: "#/components/schemas/BugReportRequest" } } }
      responses: { "201": { description: One-way report receipt }, "422": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    webCookie: { type: apiKey, in: cookie, name: vlt_web_session }
    adminCookie: { type: apiKey, in: cookie, name: vlt_admin_session }
    csrf: { type: apiKey, in: header, name: X-CSRF-Token }
    desktopBearer: { type: http, scheme: bearer, bearerFormat: Ed25519 }
    reporterBearer: { type: http, scheme: bearer, bearerFormat: opaque }
...
  responses:
    Error:
      description: Structured API error
      content: { application/json: { schema: { $ref: "#/components/schemas/APIError" } } }
...
  schemas:
...
    APIError:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        field_errors: { type: object, additionalProperties: { type: string } }
        request_id: { type: string }
...
    BugReportRequest:
      type: object
      required: [title, description]
      properties:
        title: { type: string, minLength: 3, maxLength: 160 }
        description: { type: string, minLength: 10, maxLength: 20000 }
        steps: { type: string }
        expected: { type: string }
        actual: { type: string }
        attachments:
          type: array
          maxItems: 5
          items: { type: string, format: binary }
...